requirements.txt
streamlit==1.35.0
pandas==2.2.2
scikit-learn==1.4.2
scikit-surprise==1.1.3
numpy==1.24.4
